fully automated update_stocks_daily
