+++ tradester_backend/pyproject.toml
@@ -19,6 +19,8 @@ django-cors-headers = "^3.14.0"
 requests = "^2.28.2"
 djangorestframework-simplejwt = "^5.2.2"
 python-status = "^1.0.1"
+django-crontab = "^0.7.1"
+apscheduler = "^3.10.1"
 
 [build-system]
 requires = ["poetry-core"]
